+++ example/lib/main.dart
@@ -118,11 +118,85 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 
   _buildCalendarDialogButton() {
-    final config = CalendarDatePicker2WithActionButtonsConfig(
-      calendarType: CalendarDatePicker2Type.range,
-      selectedDayHighlightColor: Colors.purple[800],
-      closeDialogOnCancelTapped: true,
+    const dayTextStyle = TextStyle(
+      color: Colors.black,
+      fontWeight: FontWeight.w700,
     );
+    final weekendTextStyle = TextStyle(
+      color: Colors.grey[500],
+      fontWeight: FontWeight.w600,
+    );
+    final anniversaryTextStyle = TextStyle(
+      color: Colors.red[400],
+      fontWeight: FontWeight.w700,
+      decoration: TextDecoration.underline,
+    );
+    final config = CalendarDatePicker2WithActionButtonsConfig(
+        dayTextStyle: dayTextStyle,
+        calendarType: CalendarDatePicker2Type.range,
+        selectedDayHighlightColor: Colors.purple[800],
+        closeDialogOnCancelTapped: true,
+        firstDayOfWeek: 1,
+        weekdayLabelTextStyle: const TextStyle(
+          color: Colors.black87,
+          fontWeight: FontWeight.bold,
+        ),
+        controlsTextStyle: const TextStyle(
+          color: Colors.black,
+          fontSize: 15,
+          fontWeight: FontWeight.bold,
+        ),
+        dayTextStylePredicate: ({required date}) {
+          TextStyle? textStyle;
+          if (date.weekday == DateTime.saturday ||
+              date.weekday == DateTime.sunday) {
+            textStyle = weekendTextStyle;
+          }
+          if (DateUtils.isSameDay(date, DateTime(2021, 1, 25))) {
+            textStyle = anniversaryTextStyle;
+          }
+          return textStyle;
+        },
+        dayBuilder: ({
+          required date,
+          textStyle,
+          decoration,
+          isSelected,
+          isDisabled,
+          isToday,
+        }) {
+          Widget? dayWidget;
+          if (date.day % 3 == 0 && date.day % 9 != 0) {
+            dayWidget = Container(
+              decoration: decoration,
+              child: Center(
+                child: Stack(
+                  alignment: AlignmentDirectional.center,
+                  children: [
+                    Text(
+                      MaterialLocalizations.of(context).formatDecimal(date.day),
+                      style: textStyle,
+                    ),
+                    Padding(
+                      padding: const EdgeInsets.only(top: 27.5),
+                      child: Container(
+                        height: 4,
+                        width: 4,
+                        decoration: BoxDecoration(
+                          borderRadius: BorderRadius.circular(5),
+                          color: isSelected == true
+                              ? Colors.white
+                              : Colors.grey[500],
+                        ),
+                      ),
+                    ),
+                  ],
+                ),
+              ),
+            );
+          }
+          return dayWidget;
+        });
     return Padding(
       padding: const EdgeInsets.all(15),
       child: Row(
@@ -137,10 +211,6 @@ class _MyHomePageState extends State<MyHomePage> {
                 borderRadius: BorderRadius.circular(15),
                 initialValue: _dialogCalendarPickerValue,
                 dialogBackgroundColor: Colors.white,
-                selectableDayPredicate: (day) => !day
-                    .difference(_dialogCalendarPickerValue[0]!
-                        .subtract(const Duration(days: 5)))
-                    .isNegative,
               );
               if (values != null) {
                 // ignore: avoid_print
@@ -182,6 +252,9 @@ class _MyHomePageState extends State<MyHomePage> {
       disabledDayTextStyle: const TextStyle(
         color: Colors.grey,
       ),
+      selectableDayPredicate: (day) => !day
+          .difference(DateTime.now().subtract(const Duration(days: 3)))
+          .isNegative,
     );
     return Column(
       mainAxisSize: MainAxisSize.min,
@@ -193,9 +266,6 @@ class _MyHomePageState extends State<MyHomePage> {
           initialValue: _singleDatePickerValueWithDefaultValue,
           onValueChanged: (values) =>
               setState(() => _singleDatePickerValueWithDefaultValue = values),
-          selectableDayPredicate: (day) => !day
-              .difference(DateTime.now().subtract(const Duration(days: 3)))
-              .isNegative,
         ),
         const SizedBox(height: 10),
         Row(